feat: optimize photo load rerender
